+++ packages/pedometer/example/ios/Flutter/flutter_export_environment.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/Users/bardram/dev/flutter"
-export "FLUTTER_APPLICATION_PATH=/Users/bardram/dev/flutter-plugins/packages/pedometer/example"
+export "FLUTTER_ROOT=/opt/homebrew/Caskroom/flutter/3.24.3/flutter"
+export "FLUTTER_APPLICATION_PATH=/Users/arata/Developer/carp/flutter-plugins/packages/pedometer/example"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"